@@ -7,16 +7,19 @@ import (
 	"reflect"
 	"testing"
 
-	"github.com/jackc/pgx/v4/pgxpool"
-	"github.com/stretchr/testify/assert"
+	crud "github.com/luanapp/gin-example/pkg/crud/mocks"
 
 	"github.com/gin-gonic/gin"
 	"github.com/google/uuid"
+	"github.com/jackc/pgx/v5/pgxpool"
+	"github.com/stretchr/testify/assert"
+
+	"github.com/luanapp/gin-example/pkg/model"
 )
 
 var (
 	id, _       = uuid.NewUUID()
-	mockSpecies = []Species{
+	mockSpecies = []model.Species{
 		{Id: id.String(), ScientificName: "Phyllobates terribilis", Genus: "Phyllobates", Family: "Dendrobatidae", Order: "Anura", Class: "Amphibia", Phylum: "Chordata", Kingdom: "Animalia"},
 		{Id: id.String(), ScientificName: "Allobates alessandroi", Genus: "Allobates", Family: "Dendrobatidae", Order: "Anura", Class: "Amphibia", Phylum: "Chordata", Kingdom: "Animalia"},
 		{Id: id.String(), ScientificName: "Ameerega bilinguis", Genus: "Ameerega", Family: "Dendrobatidae", Order: "Anura", Class: "Amphibia", Phylum: "Chordata", Kingdom: "Animalia"},
@@ -26,19 +29,19 @@ var (
 )
 
 func TestNewHandler(t *testing.T) {
-	repo := &repository{conn: &pgxpool.Pool{}}
-	got := NewHandler(repo)
-	expected := &Handler{repo}
+	repo := &repository[model.Species]{conn: &pgxpool.Pool{}}
+	got := NewHandler[model.Species](repo)
+	expected := &Handler[model.Species]{repo}
 	assert.True(t, reflect.DeepEqual(got, expected))
 }
 
 func TestHandler_GetAll_StatusOK(t *testing.T) {
 	// Arrange
-	repo := &repositoryMock{}
-	repo.On("getAll").Return(mockSpecies, nil)
+	repo := crud.NewMockRepository[model.Species](t)
+	repo.EXPECT().GetAll().Return(mockSpecies, nil)
 
 	e := gin.Default()
-	handler := NewHandler(repo)
+	handler := NewHandler[model.Species](repo)
 	e.GET("/species", handler.GetAll)
 
 	req := httptest.NewRequest(http.MethodGet, "/species", nil)
@@ -54,11 +57,11 @@ func TestHandler_GetAll_StatusOK(t *testing.T) {
 
 func TestHandler_GetAll_Status500(t *testing.T) {
 	// Arrange
-	repo := &repositoryMock{}
-	repo.On("getAll").Return([]Species{}, errors.New("some unexpected error"))
+	repo := crud.NewMockRepository[model.Species](t)
+	repo.EXPECT().GetAll().Return([]model.Species{}, errors.New("some unexpected error"))
 
 	e := gin.Default()
-	handler := NewHandler(repo)
+	handler := NewHandler[model.Species](repo)
 	e.GET("/species", handler.GetAll)
 
 	req := httptest.NewRequest(http.MethodGet, "/species", nil)
