@@ -3,13 +3,23 @@ package logger
 import (
 	"os"
 
+	"github.com/luanapp/gin-example/config/env"
+
 	"go.uber.org/zap"
+	"go.uber.org/zap/zapcore"
+)
+
+var (
+	highPriority = zap.LevelEnablerFunc(func(lvl zapcore.Level) bool {
+		return lvl >= zapcore.ErrorLevel
+	})
 )
 
 func NewLogger() (l *zap.Logger) {
-	env := os.Getenv("USING_GIN_ENV")
-	if env == "production" {
-		l, _ = zap.NewProduction(zap.AddCaller())
+	if env.IsProduction() {
+		jsonEncoder := zapcore.NewJSONEncoder(zap.NewProductionEncoderConfig())
+		core := zapcore.NewCore(jsonEncoder, zapcore.AddSync(os.Stdout), highPriority)
+		l = zap.New(core, zap.AddCaller())
 	} else {
 		l, _ = zap.NewDevelopment(zap.AddCaller())
 	}
